requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyascii==0.3.2
asttokens==2.4.1
astunparse==1.6.3
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
contractions==0.1.73
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
emoji==2.10.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
gensim==4.3.2
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
h5py==3.10.0
httplib2==0.22.0
huggingface-hub==0.20.3
idna==3.6
ipykernel==6.29.2
ipython==8.22.1
itsdangerous==2.1.2
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jwt==1.3.1
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
panda==0.3.1
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
pure-eval==0.2.2
pyahocorasick==2.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
segtok==1.5.11
sentence-transformers==2.5.1
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
textsearch==0.0.24
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
typer==0.9.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wordcloud==1.9.3
wrapt==1.14.1
yake==0.4.8
